app/requirements.txt
gradio>=4.0.0
onnxruntime-gpu
opencv-python-headless
numpy
Pillow
